feat: Refactor QueryBuilder UI with new Header, ControlPanel, and VisualBuilder components
